Fixing issues with Glance image upload function
On first run, multiple images were being uploaded with
the same name.  This patch limits the task to the first
Magnum container to prevent this.  Additionally, the Glance
client was not installed on the server, thus resulting
in an error when attemping to upload the image.  This is
now included in the magnum_requires_pip_packages variable.

Change-Id: I03183881390fe98b492b010b7627d5a444a7d7b3