Only install to virtual environment
Remove all tasks and variables related to toggling between installation
of magnum inside or outside of a Python virtual environment.
Installing within a venv is now the only supported deployment.

Additionally, a few changes have been made to make the creation of the
venv more resistant to interruptions during a run of the role.
* unarchiving a pre-built venv will now also occur when the venv
  directory is created, not only after being downloaded
* virtualenv-tools is run against both pre-built and non pre-built venvs
  to account for interruptions during or prior to unarchiving

Change-Id: Ida3320575e628076a1e9c99ba1b2bb04e7c3c4d8
Implements: blueprint only-install-venvs