Force Ansible to use dynamic includes
Ansible 2.1.1 introduces a regression in the way conditional
includes are handled which results in every task in the
included file being evaluated even if the condition for the
include is not met. This extends the run time significantly
for a deployment.

This patch forces all conditional includes to be dynamic.

Change-Id: I508c83ee2ba774e87105c72a10be899bdc2a71b9
Related-Bug: https://github.com/ansible/ansible/issues/17687